refactor(bridge): extract StreamMessageHandler
