+++ Sources/Ticker/App/Bridge/StreamMessageHandler.swift
@@ -1,0 +1,200 @@
+import AppKit
+import Foundation
+
+protocol StreamMessageHandlerDelegate: AnyObject {
+    func setCurrentStreamIdForFileDrops(_ streamId: UUID?)
+    func clearCurrentStreamIdForFileDrops(ifMatches streamId: UUID)
+    func closePDFPaneIfShowingDifferentStream(_ streamId: UUID) async
+}
+
+final class StreamMessageHandler: BridgeMessageHandler {
+    let handledTypes: Set<String> = [
+        "loadStreams",
+        "loadStream",
+        "createStream",
+        "updateStreamTitle",
+        "deleteStream",
+        "saveStreamDocument",
+        "exportStream"
+    ]
+
+    private let persistence: PersistenceService
+    private let bridgeService: BridgeService
+    private let assetService: AssetService
+    private weak var delegate: StreamMessageHandlerDelegate?
+
+    init?(container: ServiceContainer, delegate: StreamMessageHandlerDelegate) {
+        guard let persistence = container.persistence else { return nil }
+        self.persistence = persistence
+        self.bridgeService = container.bridgeService
+        self.assetService = container.assetService
+        self.delegate = delegate
+    }
+
+    func handle(_ message: BridgeMessage) async {
+        switch message.type {
+        case "loadStreams":
+            do {
+                let summaries = try persistence.loadStreamSummaries()
+                let payload = StreamCodec.encodeSummaries(summaries)
+                bridgeService.send(BridgeMessage(type: "streamsLoaded", payload: payload))
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to load streams (\(DebugLog.errorSummary(error)))")
+            }
+
+        case "loadStream":
+            guard let payload = message.payload,
+                  let idValue = payload["id"]?.value as? String,
+                  let id = UUID(uuidString: idValue) else {
+                DebugLog.log("[WebViewManager] Invalid loadStream payload")
+                return
+            }
+            delegate?.setCurrentStreamIdForFileDrops(id)
+            await delegate?.closePDFPaneIfShowingDifferentStream(id)
+            do {
+                if let stream = try persistence.loadStream(id: id) {
+                    let document = try persistence.loadOrCreateStreamDocument(streamId: id)
+
+                    let streamPayload = StreamCodec.encodeStream(stream, document: document)
+                    bridgeService.send(BridgeMessage(type: "streamLoaded", payload: ["stream": AnyCodable(streamPayload)]))
+                }
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to load stream (\(DebugLog.errorSummary(error)))")
+            }
+
+        case "createStream":
+            let title = (message.payload?["title"]?.value as? String) ?? "Untitled"
+            do {
+                let stream = try persistence.createStream(title: title)
+                delegate?.setCurrentStreamIdForFileDrops(stream.id)
+                let document = try persistence.loadOrCreateStreamDocument(streamId: stream.id)
+                let streamPayload = StreamCodec.encodeStream(stream, document: document)
+                bridgeService.send(BridgeMessage(type: "streamLoaded", payload: ["stream": AnyCodable(streamPayload)]))
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to create stream (\(DebugLog.errorSummary(error)))")
+            }
+
+        case "updateStreamTitle":
+            guard let payload = message.payload,
+                  let idValue = payload["id"]?.value as? String,
+                  let id = UUID(uuidString: idValue),
+                  let title = payload["title"]?.value as? String else {
+                DebugLog.log("[WebViewManager] Invalid updateStreamTitle payload")
+                return
+            }
+            do {
+                if var stream = try persistence.loadStream(id: id) {
+                    stream.title = title
+                    try persistence.updateStream(stream)
+                    bridgeService.send(BridgeMessage(type: "streamTitleUpdated", payload: ["id": AnyCodable(id.uuidString), "title": AnyCodable(title)]))
+                }
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to update stream title (\(DebugLog.errorSummary(error)))")
+            }
+
+        case "deleteStream":
+            guard let payload = message.payload,
+                  let idValue = payload["id"]?.value as? String,
+                  let id = UUID(uuidString: idValue) else {
+                DebugLog.log("[WebViewManager] Invalid deleteStream payload")
+                return
+            }
+            do {
+                try persistence.deleteStream(id: id)
+                delegate?.clearCurrentStreamIdForFileDrops(ifMatches: id)
+                // Also delete stream assets (images, etc.)
+                try? assetService.deleteAssets(for: id)
+                // Reload streams list
+                let summaries = try persistence.loadStreamSummaries()
+                let summariesPayload = StreamCodec.encodeSummaries(summaries)
+                bridgeService.send(BridgeMessage(type: "streamsLoaded", payload: summariesPayload))
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to delete stream (\(DebugLog.errorSummary(error)))")
+            }
+
+        case "saveStreamDocument":
+            guard let payload = message.payload,
+                  let streamIdValue = payload["streamId"]?.value as? String,
+                  let streamId = UUID(uuidString: streamIdValue),
+                  let markdown = payload["markdown"]?.value as? String else {
+                DebugLog.log("[WebViewManager] Invalid saveStreamDocument payload")
+                return
+            }
+            do {
+                try persistence.saveStreamDocument(streamId: streamId, markdown: markdown)
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to save stream document (\(DebugLog.errorSummary(error)))")
+            }
+
+        case "exportStream":
+            guard let payload = message.payload,
+                  let streamIdString = payload["streamId"]?.value as? String,
+                  let streamId = UUID(uuidString: streamIdString),
+                  let format = payload["format"]?.value as? String else {
+                DebugLog.log("[WebViewManager] Invalid exportStream payload")
+                return
+            }
+
+            do {
+                guard let stream = try persistence.loadStream(id: streamId) else {
+                    DebugLog.log("[WebViewManager] Stream not found for export")
+                    bridgeService.send(BridgeMessage(type: "exportError", payload: [
+                        "streamId": AnyCodable(streamIdString),
+                        "error": AnyCodable("Stream not found")
+                    ]))
+                    return
+                }
+                let document = try persistence.loadOrCreateStreamDocument(streamId: streamId)
+
+                // Convert to export format
+                let content = StreamCodec.formatStreamForExport(stream: stream, document: document, format: format)
+                let fileExtension = format == "markdown" ? ".md" : ".txt"
+                let suggestedName = StreamCodec.sanitizeFilename(stream.title) + fileExtension
+
+                // Show save panel on main thread
+                await MainActor.run {
+                    let savePanel = NSSavePanel()
+                    savePanel.nameFieldStringValue = suggestedName
+                    // Use appropriate content type for the format
+                    if format == "markdown" {
+                        savePanel.allowedContentTypes = [.init(filenameExtension: "md") ?? .plainText]
+                    } else {
+                        savePanel.allowedContentTypes = [.plainText]
+                    }
+                    savePanel.message = "Export stream as \(format == "markdown" ? "Markdown" : "Plain Text")"
+
+                    let result = savePanel.runModal()
+                    if result == .OK, let url = savePanel.url {
+                        do {
+                            try content.write(to: url, atomically: true, encoding: .utf8)
+                            bridgeService.send(BridgeMessage(type: "exportComplete", payload: [
+                                "streamId": AnyCodable(streamId.uuidString),
+                                "path": AnyCodable(url.path)
+                            ]))
+                        } catch {
+                            DebugLog.log("[WebViewManager] Failed to write export file (\(DebugLog.errorSummary(error)))")
+                            bridgeService.send(BridgeMessage(type: "exportError", payload: [
+                                "streamId": AnyCodable(streamId.uuidString),
+                                "error": AnyCodable(error.localizedDescription)
+                            ]))
+                        }
+                    } else {
+                        // User canceled - no error, just inform frontend
+                        bridgeService.send(BridgeMessage(type: "exportCanceled", payload: [
+                            "streamId": AnyCodable(streamId.uuidString)
+                        ]))
+                    }
+                }
+            } catch {
+                DebugLog.log("[WebViewManager] Failed to load stream for export (\(DebugLog.errorSummary(error)))")
+                bridgeService.send(BridgeMessage(type: "exportError", payload: [
+                    "streamId": AnyCodable(streamIdString),
+                    "error": AnyCodable(error.localizedDescription)
+                ]))
+            }
+
+        default:
+            DebugLog.log("[StreamMessageHandler] Unknown message type: \(message.type)")
+        }
+    }
+}
